refactor: Moved commons utilities to commons.py
