+++ modelfiles/qwen
@@ -1,4 +1,4 @@
-FROM mlx-community/Qwen3.5-9B-MLX-4bit
+FROM mlx-community/Qwen3.5-4B-MLX-4bit
 SYSTEM """
 You are Tiles, a local-first private AI assistant.
